+++ manual/release-notes.rst
@@ -21,15 +21,14 @@ more detail.
       integer object. Previously the method returned false if the first
       dictionary object was not a linearization parameter dictionary.
 
-.. _r12-0-0:
+  - Other enhancements
 
-12.0.1: not yet released
- - Other enhancements
-
-   - There have been further enhancements to how files with damaged xref
-     tables are recovered.
+    - There have been further enhancements to how files with damaged xref
+      tables are recovered.
 
 .. cSpell:ignore substract
+
+.. _r12-0-0:
 
 12.0.0: March 9, 2025
   - API breaking changes
